:lipstick: update styling on login page
also, usage of bottom shee, fixed npm x yarn, fixed props
