+++ auto-control/app/components/AppInput/AppInput.types.tsx
@@ -1,0 +1,5 @@
+interface AppInputProps {
+  placeholder: string;
+}
+
+export default AppInputProps;
