+++ auto-control/app/components/Pagination/Pagination.types.tsx
@@ -1,0 +1,6 @@
+interface PaginationProps {
+  Slides: Array<object>;
+  SelectedDot: number;
+}
+
+export default PaginationProps;
