+++ auto-control/app/components/Slider/Slider.tsx
@@ -1,39 +1,31 @@
-import React, { useCallback, useRef, useState } from "react";
+// React and React Native Imports
+import React, { useRef, useState } from "react";
+import { SafeAreaView } from "react-native-safe-area-context";
 import { FlatList, Pressable, View, Text } from "react-native";
-
-import SliderItem from "../SliderItem/SliderItem";
-
-import Slides from "../../resources/onboarding-slides";
-
-import styles from "./Slider.styles";
-import OnboardingButton from "../OnboardingButton/OnboardingButton";
-import Pagination from "../Pagination/Pagination";
-
-import RightArrow from "../../../assets/images/right-arrow.svg";
-
 import { Image } from "expo-image";
 
+// Third-Party Libraries
 import BottomSheet, { BottomSheetView } from "@gorhom/bottom-sheet";
 
-interface itemProps {
-  item: {
-    id: number;
-    name: string;
-    title: string;
-    description: string;
-    image: object;
-  };
-}
+// Project Resources
+import RightArrow from "../../../assets/images/right-arrow.svg";
+import Slides from "../../resources/onboarding-slides";
+import styles from "./Slider.styles";
+
+// Custom Components
+import SliderItem from "../SliderItem/SliderItem";
+import OnboardingButton from "../OnboardingButton/OnboardingButton";
+import Pagination from "../Pagination/Pagination";
+import Login from "@/app/screens/Login/Login";
+
+// Type Imports
+import itemProps from "./Slider.types";
 
 const Slider = () => {
   const [currentIndex, setCurrentIndex] = useState(0);
   const flatListRef = useRef<FlatList>(null);
 
   const bottomSheetRef = useRef<BottomSheet>(null);
-
-  const handleSheetChanges = useCallback((index: number) => {
-    console.log("handleSheetChanges", index);
-  }, []);
 
   const handleNextPress = () => {
     const nextIndex = currentIndex + 1;
@@ -49,48 +41,58 @@ const Slider = () => {
     setCurrentIndex(lastIndex);
   };
 
+  const handleExpandBottomSheet = () => {
+    bottomSheetRef.current?.snapToIndex(2);
+  };
+
   const snapPoints = ["25%", "50%", "90%"];
 
   return (
-    <View style={styles.container}>
-      <View style={styles.headerTextContainer}>
-        <Pressable onPress={handleGoToLastSlide}>
-          <Text style={styles.headerText}>Pular</Text>
-        </Pressable>
-        <Image source={RightArrow} style={{ width: 20, height: 20 }} />
-      </View>
-      <FlatList
-        horizontal
-        data={Slides}
-        pagingEnabled
-        ref={flatListRef}
-        scrollEnabled={false}
-        snapToAlignment="center"
-        showsHorizontalScrollIndicator={false}
-        contentContainerStyle={styles.list}
-        keyExtractor={(item) => item.id.toString()}
-        renderItem={({ item }: itemProps) => <SliderItem item={item} />}
-      />
-      <Pagination Slides={Slides} SelectedDot={currentIndex} />
-      {currentIndex !== Slides.length - 1 ? (
-        <OnboardingButton text="Próximo" onPress={handleNextPress} />
-      ) : (
-        <View style={styles.authContainer}>
-          <OnboardingButton text="Criar conta" onPress={handleNextPress} />
-          <Text style={styles.authLoginText}>Login</Text>
+    <>
+      <SafeAreaView>
+        <View style={styles.container}>
+          <View style={styles.headerTextContainer}>
+            <Pressable onPress={handleGoToLastSlide}>
+              <Text style={styles.headerText}>Pular</Text>
+            </Pressable>
+            <Image source={RightArrow} style={{ width: 20, height: 20 }} />
+          </View>
+          <FlatList
+            horizontal
+            data={Slides}
+            pagingEnabled
+            ref={flatListRef}
+            scrollEnabled={false}
+            snapToAlignment="center"
+            showsHorizontalScrollIndicator={false}
+            contentContainerStyle={styles.list}
+            keyExtractor={(item) => item.id.toString()}
+            renderItem={({ item }: itemProps) => <SliderItem item={item} />}
+          />
+          <Pagination Slides={Slides} SelectedDot={currentIndex} />
+          {currentIndex !== Slides.length - 1 ? (
+            <OnboardingButton text="Próximo" onPress={handleNextPress} />
+          ) : (
+            <View style={styles.authContainer}>
+              <OnboardingButton text="Criar conta" onPress={handleNextPress} />
+              <Pressable onPress={handleExpandBottomSheet}>
+                <Text style={styles.authLoginText}>Login</Text>
+              </Pressable>
+            </View>
+          )}
         </View>
-      )}
+      </SafeAreaView>
       <BottomSheet
         ref={bottomSheetRef}
         enablePanDownToClose
-        onChange={handleSheetChanges}
+        index={-1}
         snapPoints={snapPoints}
       >
         <BottomSheetView style={styles.contentContainer}>
-          <Text>Awesome 🎉</Text>
+          <Login />
         </BottomSheetView>
       </BottomSheet>
-    </View>
+    </>
   );
 };
 export default Slider;
