+++ auto-control/app/components/Slider/Slider.types.tsx
@@ -1,0 +1,11 @@
+interface itemProps {
+  item: {
+    id: number;
+    name: string;
+    title: string;
+    description: string;
+    image: object;
+  };
+}
+
+export default itemProps;
