+++ auto-control/app/index.tsx
@@ -1,4 +1,3 @@
-import { SafeAreaView } from "react-native";
 import Onboarding from "./screens/Onboarding/Onboarding";
 import { LinearGradient } from "expo-linear-gradient";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
@@ -7,9 +6,7 @@ const Index = () => {
   return (
     <GestureHandlerRootView>
       <LinearGradient colors={["#2282FF", "#326aee"]}>
-        <SafeAreaView>
-          <Onboarding />
-        </SafeAreaView>
+        <Onboarding />
       </LinearGradient>
     </GestureHandlerRootView>
   );
